+++ bundle/src/main/java/com/adobe/acs/commons/workflow/synthetic/impl/cq/SyntheticWorkItem.java
@@ -27,12 +27,14 @@ import com.day.cq.workflow.exec.WorkflowData;
 import com.day.cq.workflow.metadata.MetaDataMap;
 import com.day.cq.workflow.model.WorkflowNode;
 
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
 import java.util.Date;
 import java.util.Dictionary;
 import java.util.Hashtable;
 import java.util.UUID;
 
-public class SyntheticWorkItem implements WorkItem {
+public class SyntheticWorkItem implements InvocationHandler {
     private static final String CURRENT_ASSIGNEE = "Synthetic Workflow";
     private final UUID uuid = UUID.randomUUID();
     private Date timeStarted = null;
@@ -43,22 +45,50 @@ public class SyntheticWorkItem implements WorkItem {
 
     private MetaDataMap metaDataMap = new SyntheticMetaDataMap();
 
-    public SyntheticWorkItem(final WorkflowData workflowData) {
+    private SyntheticWorkItem(final WorkflowData workflowData) {
         this.workflowData = workflowData;
         this.timeStarted = new Date();
     }
 
+    public static SyntheticWorkItem createSyntheticWorkItem(WorkflowData workflowData) {
+        return new SyntheticWorkItem(workflowData);
+    }
+
     @Override
+    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+        String methodName = method.getName();
+        switch (methodName) {
+            case "getTimeStarted":
+                return getTimeStarted();
+            case "getTimeEnded":
+                return getTimeEnded();
+            case "getWorkflow":
+                return getWorkflow();
+            case "getNode":
+                return getNode();
+            case "getId":
+                return getId();
+            case "getWorkflowData":
+                return getWorkflowData();
+            case "getCurrentAssignee":
+                return getCurrentAssignee();
+            case "getMetaData":
+                return getMetaData();
+            case "getMetaDataMap":
+                return getMetaDataMap();
+            default:
+                throw new UnsupportedOperationException("CQ SYNTHETICWORKFLOW ITEM >> NO IMPLEMENTATION FOR " + methodName);
+        }
+    }
+
     public final String getId() {
         return uuid.toString() + "_" + this.getWorkflowData().getPayload();
     }
 
-    @Override
     public final Date getTimeStarted() {
         return this.timeStarted == null ? null : (Date) this.timeStarted.clone();
     }
 
-    @Override
     public final Date getTimeEnded() {
         return this.timeEnded == null ? null : (Date) this.timeEnded.clone();
     }
@@ -71,12 +101,10 @@ public class SyntheticWorkItem implements WorkItem {
         }
     }
 
-    @Override
     public final WorkflowData getWorkflowData() {
         return this.workflowData;
     }
 
-    @Override
     public final String getCurrentAssignee() {
         return CURRENT_ASSIGNEE;
     }
@@ -85,7 +113,6 @@ public class SyntheticWorkItem implements WorkItem {
      * @deprecated deprecated in interface
      */
     @Deprecated
-    @Override
     @SuppressWarnings("squid:S1149")
     public final Dictionary<String, String> getMetaData() {
         final Dictionary<String, String> dictionary = new Hashtable<String, String>();
@@ -103,24 +130,21 @@ public class SyntheticWorkItem implements WorkItem {
      *
      * @return the WorkItem's MetaDataMap
      */
-    @Override
     public final MetaDataMap getMetaDataMap() {
         return this.metaDataMap;
     }
 
-    @Override
     public final Workflow getWorkflow() {
         return this.workflow;
     }
 
-    public final void setWorkflow(final SyntheticWorkflow workflow) {
-        workflow.setActiveWorkItem(this);
+    public void setWorkflow(final WorkItem proxy, final SyntheticWorkflow workflow) {
+        workflow.setActiveWorkItem(proxy);
         this.workflow = workflow;
     }
 
     /* Unimplemented Methods */
 
-    @Override
     public final WorkflowNode getNode() {
         return null;
     }